:car: JWT 리프레시 토큰 기간 늘림

+++ src/main/java/com/dev6/rejordbe/config/JwtConfig.java
@@ -17,6 +17,6 @@ public class JwtConfig {
 
     @lombok.Data
     public static class TokenPeriod {
-        private int period;
+        private Long period;
     }
 }
